Rhiannon's Poo Predictor App Files ----------------------------------
All the files to run a sinatra app (`rhiannon.rb`) that allows you to choose from a list of foods and then directs you to a page that tells you what poo you will do. Uses the html5 bootstrap and haml for the views. Basic ruby, largely encapsulated in `rhiannon.rb` for the app.